// types/car-quote-data.ts
export type CarQuoteData = {
    vehicleNo?: string;
    nric?: string;
    email?: string;
    phoneNumber?: string;
    dateOfBirth?: string;
    vehicleMake?: string;
    vehicleModel?: string;
    registrationYear?: string;
    vehicleFinance?: string;
    drivingExperiences?: string;
    ncd?: string;
    numberOfClaims?: string;
    policyStartDate?: string;
    policyEndDate?: string;
};